refactor: 패키지 구조 변경
- request 패키지를 message 패키지로 이동
