+++ tomcat/src/main/java/org/apache/coyote/http11/message/request/RequestPath.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.request;
+package org.apache.coyote.http11.message.request;
 
 public class RequestPath {
     private final String path;
